Implement MultiImage working with Pick (MIPick)
This is the first implementation of the MultiImage and Pick plugins
working in conjunction. Currently to modified MIPick plugin needs to be
used.

To see in action, simply ensure that MultiImage is started first, then
start MIPick. Load up a couple images the overlap the same sky, and then
pick away. Take special note of what happens when one switches images.
For most dramatic results, use images of different wavelengths or
resolution.

There is much refactor that needs to occur, but functionality is there.

Resolves Issue #31